Major updates to setup and install scripts
It should now be possible to just download the install script and run it
and it should download, install and run everything needed to get going
with my dot files.  This needs further testing and documentation.
Supporting OS's and Distros currently are Macs, Ubuntu and Raspbian.  If
you want another OS/Distro support please supply a PR for your OS.