Added WinGPG connection script
This is needed to allow WSL access to Yubikeys.  I need a better
script.  One which can run unattended and doesn't generate jobs
which require terminating.  My first attempts have all failed.  I
may revisit this in the future but I've decided for now not to waste
any more time on it.