Updated the Ubuntu script to include GPG in WSL
OK this is complex and horrible as it means jumping between Bash and Powershell in odd ways but it seems to mostly work.  You have to pair it with the WinGPG script I have.  You also need to populate your instance of WinGPG with your Yubikey and such like.  I have private scripts which do this later.  This setups the ground work for doing so.