Cleaned up some stuff.  Added Razer Xorg Conf File
The big change here is addition of the Xorg Razer conf file.  If you
have to many Razer devices you can cause X to have issues.  Most Razer
devices present as a keyboard or similar input device.  To many and Xorg
crashes.  This file fixes this.  I'll update the Ubuntu setup script to
copy this to the right place.