Split out the ubuntu-setup script
There is now a system setup script which install programs and packages
for everyone and there is the ubuntu-setup script which installs
everything which needs a user to be install (e.g. changing the shell,
install streamdec, wsl stuffs, etc)

This allows me to run the setup script and clean up my system before
first boot.