Major changes - Moved to oh-my-posh and vim-plug
This lays the ground work for moving to a new dot-files repo by
switching from powerline10k to oh-my-posh.  In addition I've switched
from pathogen to vim-plug.

There's now more consistent theming across vim and the command prompted.
The base them for vim is now nord and the colours used in the oh-my-posh
theme is also the colour palette provided by nord.

Support for the raw linux console is currently misssing and files
relating to Powerline10k and Spacesship are likely to be removed in a
following commit unless I completely move to a new dot-files repo.